requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.4
async-generator==1.10
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
backports.entry-points-selectable==1.1.0
bleach==4.0.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.4
django-crispy-forms==1.12.0
django-environ==0.4.5
django-filter==2.4.0
django-heroku==0.3.1
django-phonenumber-field==5.2.0
django-six==1.0.4
djongo==1.3.6
docopt==0.6.2
drf-yasg==1.20.0
entrypoints==0.3
filelock==3.0.12
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
install==1.3.4
ipykernel==6.0.3
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
Markdown==3.3.4
MarkupSafe==2.0.1
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
packaging==21.0
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
pipreqs==0.4.10
platformdirs==2.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psycopg2==2.8.6
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.9.0
pymongo==3.12.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
pywin32==301
pywinpty==1.1.3
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.9.0
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
Send2Trash==1.8.0
six==1.16.0
sqlparse==0.2.4
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.2.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.7.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
yarg==0.1.9
